use long instead of int to avoid overflow

+++ jgiven-core/src/main/java/com/tngtech/jgiven/report/model/ReportStatistics.java
@@ -6,5 +6,5 @@ public class ReportStatistics {
     public int numCases;
     public int numFailedCases;
     public int numSteps;
-    public int durationInNanos;
+    public long durationInNanos;
 }
